+++ my_flutter/lib/src/resources/repositories/request_repository.dart
@@ -1,4 +1,3 @@
-import 'dart:ffi';
 
 import '../../models/lists/requests_list.dart';
 import '../../models/request_model.dart';
@@ -8,8 +7,8 @@ class Repository {
   final requestsApiProvider = RequestApiProvider();
 
   Future<RequestsList> fetchRequests(String type) => requestsApiProvider.fetchRequestList(type);
-  Future<RequestModel> fetchRequest(Long id) => requestsApiProvider.fetchRequest(id);
-  Future<RequestModel> deleteRequest(Long id) => requestsApiProvider.deleteRequest(id);
+  Future<RequestModel> fetchRequest(int id) => requestsApiProvider.fetchRequest(id);
+  Future<RequestModel> deleteRequest(int id) => requestsApiProvider.deleteRequest(id);
   Future<RequestModel> addRequest(RequestModel request) => requestsApiProvider.addRequest(request);
   Future<RequestModel> updateRequest(RequestModel request) => requestsApiProvider.updateRequest(request);
